requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.4
psycopg2==2.9.6
python-dotenv
gunicorn==21.1.0
